split island 3, 4 and 5 into seperate files
