+++ source/mdcii/mdcii/src/gam/island4.cpp
@@ -29,204 +29,6 @@
 #include "gam/islandhouse.hpp"
 
 
-Island5::Island5(uint8_t* data, uint32_t length, const std::string& name)
-  : name(name)
-  , files(Files::instance())
-{
-  memcpy((char*)&island, data, length);
-}
-
-Island5::Island5(const Island3& island3)
-{
-  island.islandNumber = island3.island3.inselnr;
-  island.posx = island3.island3.x_pos;
-  island.posy = island3.island3.y_pos;
-  island.width = island3.island3.breite;
-  island.height = island3.island3.hoehe;
-}
-
-Island5::Island5(const Island4& island4)
-{
-  memcpy(&island, &island4.island, sizeof(Island4Data));
-  island.fileNumber = island4.island.fileNumber;
-  island.modifiedFlag = IslandModified::True;
-  island.islandNumber = 0;
-}
-
-void Island5::setIslandNumber(uint8_t number)
-{
-  island.islandNumber = number;
-}
-
-void Island5::setIslandFile(uint16_t fileNumber)
-{
-  island.fileNumber = fileNumber;
-}
-
-void Island5::addIslandHouse(std::shared_ptr<Chunk> c)
-{
-  islandHouse.push_back(std::make_shared<IslandHouse>(c->chunk.data, c->chunk.length, c->chunk.name.c_str(), island.width, island.height));
-}
-
-void Island5::addIslandHouse(std::shared_ptr<IslandHouse> i)
-{
-  islandHouse.push_back(i);
-}
-
-std::string Island5::islandFileName(IslandSize size, uint8_t islandNumber, IslandClimate climate)
-{
-  return boost::str(boost::format("%s/%s%02d.scp") % islandClimateMap[climate] % islandSizeMap[size] % (int)islandNumber);
-}
-
-IslandClimate Island5::randomIslandClimate()
-{
-  std::time_t now = std::time(0);
-  boost::random::mt19937 gen{static_cast<std::uint32_t>(now)};
-  return static_cast<IslandClimate>(gen() % static_cast<uint32_t>(IslandClimate::Any));
-}
-
-void Island5::finalize()
-{
-  // island is unmodified, load bottom islandHouse from island file
-  if (island.modifiedFlag == IslandModified::False && islandHouse.size() <= 1)
-  {
-    // load the unmodified bottom layer from the island .scp file
-    auto islandFile = Island5::islandFileName(island.size, island.fileNumber, island.climate);
-    auto path = files->find_path_for_file(islandFile);
-    if (path == "")
-    {
-      throw("cannot find file: " + path);
-    }
-    std::vector<std::shared_ptr<Chunk>> chunks = Chunk::ReadChunks(path);
-    if (chunks[1]->chunk.name == "INSELHAUS")
-    {
-      auto i = std::make_shared<IslandHouse>(chunks[1]->chunk.data, chunks[1]->chunk.length, chunks[1]->chunk.name.c_str(), island.width, island.height);
-      islandHouse.push_back(i);
-    }
-    if (islandHouse.size() == 2)
-    {
-      finalIslandHouse.push_back(islandHouse[0]);
-      finalIslandHouse.push_back(islandHouse[1]);
-    }
-    // there is only one islandHouse chunk present, this is the bot layer
-    if (islandHouse.size() == 1)
-    {
-      finalIslandHouse.push_back(islandHouse[0]);
-      // create empty top
-      auto empty = std::make_shared<IslandHouse>("INSELHAUS", island.width, island.height);
-      finalIslandHouse.push_back(empty);
-    }
-  }
-  else
-  {
-    // the island is modified, first chunk is bottom
-    finalIslandHouse.push_back(islandHouse[0]);
-    // a possible second chunk is top
-    if (islandHouse.size() == 2)
-    {
-      finalIslandHouse.push_back(islandHouse[1]);
-    }
-    else
-    {
-      // create empty top
-      auto islandHouse = std::make_shared<IslandHouse>("INSELHAUS", island.width, island.height);
-      finalIslandHouse.push_back(islandHouse);
-    }
-  }
-  bottomLayer = finalIslandHouse[0];
-  topLayer = finalIslandHouse[1];
-}
-
-void Island5::setDeer(std::shared_ptr<Chunk> c)
-{
-  deer = Deer(c->chunk.data, c->chunk.length, "HIRSCH2");
-}
-
-IslandHouseData Island5::TerrainTile(uint8_t x, uint8_t y)
-{
-  // Check if top layer has something for the given position, fallback to bottom layer
-  auto h = finalIslandHouse[1]->Get(x, y);
-  if (h.id == 0xFFFF)
-  {
-    h = finalIslandHouse[0]->Get(x, y);
-  }
-
-  uint8_t xp = h.posx;
-  uint8_t yp = h.posy;
-  if ((yp > y) || (xp > x))
-  {
-    return h;
-  }
-
-  h = finalIslandHouse[1]->Get(x - xp, y - yp);
-  if (h.id == 0xFFFF)
-  {
-    h = finalIslandHouse[0]->Get(x - xp, y - yp);
-  }
-  h.posx = xp;
-  h.posy = yp;
-  return h;
-}
-
-
-TileGraphic Island5::GraphicIndexForTile(IslandHouseData& tile, uint32_t rotation)
-{
-  auto haeuser = Haeuser::Instance();
-  TileGraphic target;
-  if (tile.id == 0xFFFF)
-  {
-    target.index = -1;
-    target.groundHeight = 0;
-    return target;
-  }
-  auto info = haeuser->get_haus(tile.id);
-
-  if (!info || info.value()->Gfx == -1)
-  {
-    target.index = -1;
-    target.groundHeight = 0;
-    return target;
-  }
-  int index = info.value()->Gfx;
-  int directions = 1;
-  if (info.value()->Rotate > 0)
-  {
-    directions = 4;
-  }
-  int aniSteps = 1;
-  if (info.value()->AnimAnz > 0)
-  {
-    aniSteps = info.value()->AnimAnz;
-  }
-  index += info.value()->Rotate * ((rotation + tile.orientation) % directions);
-  switch (tile.orientation)
-  {
-    case 0:
-      index += tile.posy * info.value()->Size.w + tile.posx;
-      break;
-    case 1:
-      index += (info.value()->Size.h - tile.posx - 1) * info.value()->Size.w + tile.posy;
-      break;
-    case 2:
-      index += (info.value()->Size.h - tile.posy - 1) * info.value()->Size.w + (info.value()->Size.w - tile.posx - 1);
-      break;
-    case 3:
-      index += tile.posx * info.value()->Size.w + (info.value()->Size.w - tile.posy - 1);
-      break;
-    default:
-      break;
-  }
-  index += info.value()->Size.h * info.value()->Size.w * directions * (tile.animationCount % aniSteps);
-  target.index = index;
-  target.groundHeight = 0;
-  if (info.value()->Posoffs == 20)
-  {
-    target.groundHeight = info.value()->Posoffs;
-  }
-  return target;
-}
-
-
 Island4::Island4(uint8_t* data, uint32_t length, const std::string& name)
   : name(name)
   , files(Files::instance())
@@ -405,11 +207,4 @@ TileGraphic Island4::GraphicIndexForTile(IslandHouseData& tile, uint32_t rotatio
     target.groundHeight = info.value()->Posoffs;
   }
   return target;
-}
-
-
-Island3::Island3(uint8_t* data, uint32_t length, const std::string& name)
-  : name(name)
-{
-  memcpy((char*)&island3, data, length);
 }
